% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luca_Stephan_pkg.R
\name{active_set}
\alias{active_set}
\title{Determine the active set variables via lasso regression but it augments also the lasso selection.}
\usage{
active_set(i, d, p, X_index, y, z, z_a = NULL, bound = 0.5 * NROW(z))
}
\arguments{
\item{i}{determines which column of y is considered}

\item{d}{the decided lag-augmentation of the interest variables}

\item{p}{the lag length}

\item{X_index}{index of the Granger causing variables}

\item{y}{vector/matrix of dependent variable}

\item{z}{vector/matrix of regressors}

\item{z_a}{those variables that should not be penalized}

\item{bound}{upper bound on tuning parameter lambda}
}
\value{
the active set estimated via lasso using BIC as penalty
}
\description{
Determine the active set variables via lasso regression but it augments also the lasso selection.
}
