% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apf_fdrOpt.R
\name{apf_fdrOpt}
\alias{apf_fdrOpt}
\title{Wrapper around APFr}
\usage{
apf_fdrOpt(mat_input, FDR_max, APF_min, verbose = F)
}
\arguments{
\item{mat_input}{the input matrix}

\item{FDR_max}{the maximum false discovery rate desired}

\item{APF_min}{the minimum average power function desired}

\item{verbose}{logical for verbose console}
}
\value{
the desired constrained opt result
}
\description{
Wrapper around APFr \code{ apf_plot} to get the desired threshold on the p values such that the FDR is at most \code{ FDR_max}
and APF is closest to \code{ APF_min}
}
\keyword{internal}
