% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luca_Stephan_pkg.R
\name{bic}
\alias{bic}
\title{Lasso Bayesian Information Criterion}
\usage{
bic(y, x, b, df, bound = 0.5 * length(y))
}
\arguments{
\item{y}{vector of dependent variable}

\item{x}{vector or matrix of regressors}

\item{b}{lasso coefficients (taken from glmnet)}

\item{df}{lasso degrees of freedom (taken from glmnet)}

\item{bound}{upper bound on tuning parameter lambda: i.e. impose lasso to stop the selection after it selects 0.5T parameters.}
}
\value{
the estimated parameters with BIC
}
\description{
Do BIC on lasso output. Needs degrees of freedom coming from glmnet. Bound on lambda is optional,
imposed as 0.5T by default.
}
