% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lags_RV.R
\name{create_lags_RV}
\alias{create_lags_RV}
\title{Lags creation: Daily, Weekly, Monthly aggregation for Realized Volatilities}
\usage{
create_lags_RV(y, include.original = TRUE, trim = TRUE)
}
\arguments{
\item{y}{vector or matrix to lag}

\item{include.original}{logical, if TRUE the original series are left inside the output matrix.}

\item{trim}{logical, if TRUE the initial NAs due to the lag gets trimmed}
}
\value{
the lagged vector or matrix
}
\description{
Creates a matrix of order 3 containing Daily, Weekly, Monthly returns for Realized Volatilities.
Can include or exclude original series, and trim the NAs in the start of the sample. It should be used for HVARs.
}
\keyword{internal}
