% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_HVAR.R
\name{HDGC_HVAR}
\alias{HDGC_HVAR}
\title{Test Granger causality in High Dimensional Stationary Heterogeneous VARs}
\usage{
HDGC_HVAR(
  GCpair,
  data,
  log = TRUE,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{GCpair}{A named list with names GCto and GCfrom containing vectors of the relevant GC variables.}

\item{data}{the data matrix or something that can be transformed to a matrix containing realized volatilities.}

\item{log}{default is TRUE, if the realized volatilities are already log transformed then put =FALSE}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM test statistics and p-values: asymptotic, with finite sample correction and asymptotic with heteroscedasticity correction and Lasso selections are printed to the console
}
\description{
Test Granger causality in High Dimensional Stationary Heterogeneous VARs
}
\examples{
\dontrun{GCpair<-list("GCto"="Var 1", "GCfrom"="Var 5")}
\dontrun{HDGC_HVAR(GCpair, data=sample_dataset, log=T,parallel = F)}
}
