% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_HVAR_all.R
\name{HDGC_HVAR_all}
\alias{HDGC_HVAR_all}
\title{Granger Causality Network in High Dimensional HVARs}
\usage{
HDGC_HVAR_all(
  data,
  log = TRUE,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or an object that can be coerced to a matrix containing (stationary) realized volatilities}

\item{log}{default is TRUE, if the realized volatilities are already log transformed then put  to FALSE}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
A Granger causality matrix and Lasso selections are printed to the console
}
\description{
Wrapper around \code{\link{HDGC_HVAR_multiple}} which tests Granger causality from each variable to all other variables,
one by one. Can therefore be used to construct a network.
}
\examples{
\dontrun{HDGC_HVAR_all(data=sample_RV, log=TRUE, parallel = TRUE) }
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).

Corsi, Fulvio. "A simple approximate long-memory model of realized volatility." Journal of Financial Econometrics 7.2 (2009): 174-196.
}
