% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_HVAR_multiple.R
\name{HDGC_HVAR_multiple}
\alias{HDGC_HVAR_multiple}
\title{Test multiple combinations Granger causality in High Dimensional Stationary HVARs}
\usage{
HDGC_HVAR_multiple(
  data,
  GCpairs,
  log = TRUE,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or something that can be coerced to a matrix containing realized volatilities.}

\item{GCpairs}{it should contain a nested list: the outer list is all the pairs to be considered,
the inner list contains the GCto and GCfrom vectors needed for \code{ \link{HDGC_HVAR}}. See Example.}

\item{log}{default is TRUE, if the realized volatilities are already log transformed then put to FALSE}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{numberr of cores to use in parallel computing, default is all but one}
}
\value{
LM test statistics, p-values: asymptotic and with finite sample correction and Lasso selections are printed to the console
}
\description{
This function is a wrapper around \code{\link{HDGC_HVAR}} that allows for multiple combinations to be tested
}
\examples{
\dontrun{GC<-list(list("GCto"="Var 1", "GCfrom"="Var 2"),list("GCto"="Var 2", "GCfrom"="Var 3"))}
\dontrun{HDGC_HVAR_multiple(sample_RV,GCpairs=GC,log=TRUE)
}
}
