% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_HVAR_multiple_pairs.R
\name{HDGC_HVAR_multiple_pairs}
\alias{HDGC_HVAR_multiple_pairs}
\title{Test multiple pairs Granger causality in High Dimensional HVARs}
\usage{
HDGC_HVAR_multiple_pairs(
  data,
  GCpairs = NULL,
  GCto = NULL,
  GCfrom = NULL,
  log = TRUE,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or something that can be coerced to a matrix containing (stationary) realized volatilities}

\item{GCpairs}{it should contain a nested list. The outer list is all the pairs to be considered.
The inner list contains the GCto and GCfrom vectors needed for \code{\link{HDGC_HVAR}}. See \code{ \link{HDGC_HVAR_multiple}}.}

\item{GCto}{all combination variables Granger caused}

\item{GCfrom}{all combination variables Granger causing}

\item{log}{default is TRUE, if the realized volatilities are already log transformed then put  to FALSE}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
Granger causality matrix and Lasso selections are printed to the console
}
\description{
A wrapper around \code{\link{HDGC_HVAR_multiple}}. If GCpairs is used,
the function is the same as \code{\link{HDGC_HVAR_multiple}}. Alternatively, if a we want to test all combinations between
variables in GCto and GCfrom, these can be given directly. See Example.
}
\examples{
\dontrun{GCto = list(c("Var 1", "Var 2")); GCfrom = list(c("Var 3", "Var 4", "Var 5"))}
\dontrun{HDGC_HVAR_multiple_pairs(sample_RV,GCto,GCfrom, log=TRUE)}
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).

Corsi, Fulvio. "A simple approximate long-memory model of realized volatility." Journal of Financial Econometrics 7.2 (2009): 174-196.
}
