% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_HVAR_RV_RCoV_all.R
\name{HDGC_HVAR_RV_RCoV_all}
\alias{HDGC_HVAR_RV_RCoV_all}
\title{Networks of Realized Volatilities conditional on the set of Realized Correlations}
\usage{
HDGC_HVAR_RV_RCoV_all(
  realized_variances,
  realized_covariances,
  fisher_transf = TRUE,
  log = TRUE,
  bound = 0.5 * nrow(realized_variances),
  parallel = FALSE,
  n_cores = NULL,
  progress_bar = TRUE,
  store_selections = FALSE
)
}
\arguments{
\item{realized_variances}{Dataset of (stationary) realized volatilities. A matrix or object that can be coerced to a matrix.}

\item{realized_covariances}{Dataset of (stationary) realized covariances. A matrix or object that can be coerced to a matrix. Note: the columns should exactly
be (((ncol(realized_volatilities)^2)-ncol(realized_volatilities))/2)}

\item{fisher_transf}{Logical: if TRUE the correlations are computed and Fisher transformed}

\item{log}{Default is TRUE, if the realized volatilities are already log transformed then put to FALSE}

\item{bound}{Lower bound on lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{Nr of cores to use in parallel computing, default is all but one}

\item{progress_bar}{display a progress bar, default is true}

\item{store_selections}{store the lasso-selected variables, default is false (object can become huge in large systems)}
}
\value{
LM Chi-square test statistics (asymptotic), LM F-stats with finite sample correction, LM Chi-square (asymptotic) with heteroscedasticity correction, all with their corresponding p-value.
Lasso selections are also given if `store_selections = TRUE` (`NULL` otherwise).
}
\description{
Networks of Realized Volatilities conditional on the set of Realized Correlations
}
\examples{
\dontrun{ HDGC_HVAR_RV_RCoV_all(real_var, real_cov, fisher_transf=T, log=TRUE ,parallel = TRUE) }
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).

Corsi, Fulvio. "A simple approximate long-memory model of realized volatility." Journal of Financial Econometrics 7.2 (2009): 174-196.
}
