% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_HVAR_RVCOV.R
\name{HDGC_HVAR_RVCOV}
\alias{HDGC_HVAR_RVCOV}
\title{Test Granger causality for Realized Volatilities in High Dimensional Stationary Heterogeneous VARs conditioning on Realized Correlations}
\usage{
HDGC_HVAR_RVCOV(
  GCpair,
  realized_variances,
  realized_correlations,
  bound = 0.5 * nrow(realized_variances),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{GCpair}{A named list with names GCto and GCfrom containing vectors of the relevant GC variables.}

\item{realized_variances}{Dataset of realized volatilities. A matrix or something that can be coerced to a matrix. Note: the volatilities must not be in logs.}

\item{realized_correlations}{Dataset of realized correlations. To compute realized correlations from realized variances and realized covariances use \code{\link{Realized_corr}}}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM test statistics and p-values: asymptotic, with finite sample correction and asymptotic with heteroscedasticity correction and Lasso selections are printed to the console
}
\description{
Test Granger causality for Realized Volatilities in High Dimensional Stationary Heterogeneous VARs conditioning on Realized Correlations
}
\examples{
\dontrun{GCpair<-list("GCto"="X", "GCfrom"="Z")
HDGC_HVAR_RVCOV(GCpair, real_var, real_corr,parallel = T)}
}
