% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_VAR_all.R
\name{HDGC_VAR_all}
\alias{HDGC_VAR_all}
\title{Granger Causality Network in High Dimensional mixed Integrated and Cointegrated VARs}
\usage{
HDGC_VAR_all(
  data,
  p = 1,
  d = 0,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL,
  progress_bar = TRUE,
  store_selections = FALSE
)
}
\arguments{
\item{data}{the data matrix or object that can be coerced to a matrix.}

\item{p}{lag length of VAR}

\item{d}{order of lag augmentation corresponding to suspected max order of integration}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}

\item{progress_bar}{display a progress bar, default is true}

\item{store_selections}{store the lasso-selected variables, default is false (object can become huge in large systems)}
}
\value{
LM Chi-square test statistics (asymptotic) and LM F-stats with finite sample correction, with their corresponding p-value.
Lasso selections are also given if `store_selections = TRUE` (`NULL` otherwise).
}
\description{
Wrapper around \code{\link{HDGC_VAR_multiple}} which tests Granger causality from each variable to all other variables,
one by one. Can therefore be used to construct a network.
}
\examples{
\dontrun{HDGC_VAR_all(data=sample_dataset_I1, p=2, d=2, parallel=TRUE )}
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Inference in Non Stationary High Dimensional VARs" (2020, check the latest version at https://sites.google.com/view/luca-margaritella )

Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).
}
