% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_VAR_multiple.R, R/Luca_Stephan_pkg.R
\name{HDGC_VAR_multiple}
\alias{HDGC_VAR_multiple}
\title{Test multiple combinations Granger causality in High Dimensional mixed Integrated and Cointegrated VARs}
\usage{
HDGC_VAR_multiple(
  data,
  GCpairs,
  p = 1,
  d = 0,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)

HDGC_VAR_multiple(
  data,
  GCpairs,
  p = 1,
  d = 0,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or something that can be coerced to a matrix.}

\item{GCpairs}{it should contain a nested list. The outer list is all the pairs to be considered.
The inner list contains the GCto and GCfrom vectors needed for \code{\link{HDGC_VAR}}.}

\item{p}{lag length of the VAR}

\item{d}{order of lag augmentation corresponding to suspected max order of integration}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM test statistics, p-values: asymptotic and with finite sample correction and Lasso selections are printed to the console

LM test statistics, p-values: asymptotic and with finite sample correction and Lasso selections are printed to the console
}
\description{
This function is a wrapper around \code{\link{HDGC_VAR}} that allows for multiple combinations to be tested

This function is a wrapper around \code{\link{HDGC_VAR}} that allows for multiple combinations to be tested
}
\examples{
\dontrun{GCpairs<-list(list("GCto"="X1", "GCfrom"="X2"),list("GCto"="X2", "GCfrom"="X3"))}
\dontrun{HDGC_VAR_multiple(data, GCpairs, p=1, d=2, parallel=T)}
\dontrun{GCpairs<-list(list("GCto"="X1", "GCfrom"="X2"),list("GCto"="X2", "GCfrom"="X3"))}
\dontrun{HDGC_VAR_multiple(data, GCpairs, p=1, d=2, parallel=T)}
}
