% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_matrix_sqrt.R
\name{inv_matrix_sqrt}
\alias{inv_matrix_sqrt}
\title{Inverse square root matrix}
\usage{
inv_matrix_sqrt(M)
}
\arguments{
\item{M}{the matrix}
}
\value{
the inverse matrix square rooted
}
\description{
Inverse square root matrix
}
\keyword{internal}
