% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luca_Stephan_pkg.R
\name{lags_upbound_BIC}
\alias{lags_upbound_BIC}
\title{Lag length Selection via BIC empirical upper bound}
\usage{
lags_upbound_BIC(data, p_max = 10)
}
\arguments{
\item{data}{a dataframe or matrix of the original set of time series forming the VAR}

\item{p_max}{maximum lag length to consider, default is 10}
}
\value{
returns the estimated lag length upper bound
}
\description{
Selects the lag length p of the VAR using an empirical upper bound: residuals of
the diagonalized VAR are used to build the empirical covariance matrix and an approximation of
its determinant that uses the matrix trace is employed to be able to select p using Bayesian Information Criterion
}
\examples{
\dontrun{ lags_upbound_BIC(data,p_max=10)}
}
