% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_test.R
\name{LM_test}
\alias{LM_test}
\title{LM test}
\usage{
LM_test(y, x, z, I = 1)
}
\arguments{
\item{y}{dependent variable: Granger-caused}

\item{x}{Granger causing variable, all its lags}

\item{z}{conditioning set of variables}

\item{I}{the numberr of dependent variables, that is number of equations in the VAR.}
}
\value{
test statistics, both asymptotic chi square and finite sample corrected F and relative p values
}
\description{
Perform the LM test of significance of x for y given z.
}
\keyword{internal}
