% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_test_robust.R, R/Luca_Stephan_pkg.R
\name{LM_test_robust}
\alias{LM_test_robust}
\title{LM test heteroscedastic robust}
\usage{
LM_test_robust(y, x, z, I = 1)

LM_test_robust(y, x, z, I = 1)
}
\arguments{
\item{y}{dependent variable: Granger-caused}

\item{x}{Granger causing variable, all its lags}

\item{z}{conditioning set of variables}

\item{I}{the numberr of dependent variables, that is number of equations in the VAR.}
}
\value{
test statistics, both asymptotic chi square, finite sample corrected F, asymptotic chi square robust to heteroscedasticity and relative p values

test statistics, both asymptotic chi square, finite sample corrected F, asymptotic chi square robust to heteroscedasticity and relative p values
}
\description{
Perform the LM test of significance of x for y given z and add the robust asymptotic test statistic.

Perform the LM test of significance of x for y given z and add the robust asymptotic test statistic.
}
