% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luca_Stephan_pkg.R
\name{ols_no_int}
\alias{ols_no_int}
\title{Ordinary Least Squares without intercept}
\usage{
ols_no_int(y, x)
}
\arguments{
\item{y}{vector dependent variable}

\item{x}{a vector or matrix of regressors}
}
\value{
coefficients, fitted values, t statistics, p values, residuals
}
\description{
Perform least squares of y on x. Returns a list with the coefficients, t statistics, p values and
residual vector.
}
