% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GC_all.R
\name{Plot_GC_all}
\alias{Plot_GC_all}
\title{Plot High-Dimensional Granger causality Networks}
\usage{
Plot_GC_all(
  Comb,
  Stat_type = "FS_cor",
  alpha = 0.01,
  multip_corr = list(F, "bonferroni", gamm = c(1e-04, 0.1, 0.001), fdr.apf = c(0.05,
    0.6), verb = F),
  ...,
  cluster = list(F, 10, "black", 0.51, 1, 0)
)
}
\arguments{
\item{Comb}{result of function \code{\link{HDGC_VAR_all_I0}}, \code{\link{HDGC_VAR_multiple_pairs_I0}}, \code{\link{HDGC_VAR_all}}, \code{\link{HDGC_VAR_multiple_pairs}}, \code{\link{HDGC_HVAR_all}} or \code{\link{HDGC_HVAR_multiple_pairs}}}

\item{Stat_type}{either FS_cor (default), Asymp or Asymp_Robust respectively for F small sample correction, standard chi square test, standard chi square test with heteroscedasticity correction}

\item{alpha}{the desired probability of type one error, default is 0.01.}

\item{multip_corr}{A list: first element is logical, if TRUE a multiple testing correction using stats::p.adjust() is used. The second
element of the list define the p.adjust.method=c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", "none")). If the second element gets the name "APF_FDR" then APFr::apf_fdr which uses empirical Bayes is called and a third and fourth
elements in the mutip_corr list are required: gamm=c(,,) requires a min, max and step length values to be set for the threshold on the p_values,
fdr.apf=c(,) requires one or two values: either (NULL,value) or (value,NULL) if one wants to have specified amount of average power (fdr) no matter fdr (average power).
If both (value,value) are given, the calculated threshold will find the closest combination to both apf and fdr desired. The last element
of the list is logical: verbose=TRUE if one wants to know how much apf/fdr the testing has.}

\item{...}{all parameters for the network plot: see example and igraph documentation.}

\item{cluster}{A list: first element is logical, if TRUE a cluster plot using igraph::cluster_edge_betweenness() is plotted.
Other elements are respectively: vertex.size, vertex.label.color,vertex.label.cex, vertex.label.dist, edge.curved (see igraph for details).}
}
\value{
a network compatible with igraph
}
\description{
Plot High-Dimensional Granger causality Networks
}
\examples{
\dontrun{Plot_GC_all(Comb, "FS_cor",alpha=0.01,multip_corr=list(F), directed=T, layout.circle}
}
