% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Realized_corr.R
\name{Realized_corr}
\alias{Realized_corr}
\title{Computing Realized Correlation}
\usage{
Realized_corr(realized_variances, realized_covariances, fisher_transf = T)
}
\arguments{
\item{realized_variances}{Dataset of realized volatilities. A matrix or something that can be coerced to a matrix. Note: the volatilities must not be in logs.}

\item{realized_covariances}{Dataset of realized covariances. A matrix or something that can be coerced to a matrix.}

\item{fisher_transf}{Logical: if TRUE the correlations are computed and Fisher-transformed}
}
\value{
LM test statistics and p-values: asymptotic, with finite sample correction and asymptotic with heteroscedasticity correction and Lasso selections are printed to the console
}
\description{
This function computes realized correlations from realized variances and covariances with the possibility of
Fisher-transforming the realized correlations.
}
\examples{
\dontrun{Realized_corr(real_var, real_cov, fisher_transf=T)}
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).
}
