% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luca_Stephan_pkg.R, R/resid_covariance.R
\name{resid_covariance}
\alias{resid_covariance}
\title{Residuals Covariance Matrix}
\usage{
resid_covariance(data_array, k)

resid_covariance(data_array, k)
}
\arguments{
\item{data_array}{the array containing the different matrices}

\item{k}{the last column considered for calculating residuals}
}
\value{
the residual covariance matrix

the residual covariance matrix
}
\description{
Given an array of matrices, it calculates the residual covariance matrix between column i=1,...,K
of the first matrix of the array against column(s) $i=1,...,K of 2,..,k>=2$ matrices of the same array.

Given an array of matrices, it calculates the residual covariance matrix between column i=1,...,K
of the first matrix of the array against column(s) $i=1,...,K of 2,..,k>=2$ matrices of the same array.
}
