% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R
\docType{data}
\name{sample_dataset}
\alias{sample_dataset}
\title{Sample Data for Examples}
\format{
A matrix of 30 columns and 200 rows where each column correspond to a single non-stationary time series.
\describe{
\item{T_}{The sample size}
\item{g}{The number of covariates}
}
}
\usage{
sample_dataset
}
\description{
A dataset of 30 non-stationary time series of sample T=200 obtined by reverting a 30 stationary time series obtained through a non-sparse VAR(1) DGP.
This means effectively that the series are obtained from a VAR(2) in levels. All series have a unit root.
}
\keyword{datasets}
