% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset_I0.R
\docType{data}
\name{sample_dataset_I0}
\alias{sample_dataset_I0}
\title{Stationary Sample Data for Examples}
\format{
A matrix of 30 columns and 200 rows where each column correspond to a single stationary time series.
\describe{
\item{T_}{The sample size}
\item{g}{The number of covariates}
}
}
\usage{
sample_dataset_I0
}
\description{
A dataset of 30 stationary time series of sample T=200 obtained through a non-sparse VAR(1) DGP.
}
\keyword{datasets}
