% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_list.R
\name{simplify_list}
\alias{simplify_list}
\title{simplify output}
\usage{
simplify_list(test_list, GCpairs)
}
\arguments{
\item{test_list}{the test list from \code{ \link{LM_test}}.}

\item{GCpairs}{the pairs}
}
\value{
the statistics and pvalues
}
\description{
This function transforms and simplifies output of the GC testing functions below.
}
\keyword{internal}
