% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_matrix.R
\name{split_matrix}
\alias{split_matrix}
\title{Split Big Matrix}
\usage{
split_matrix(M, r, c)
}
\arguments{
\item{M}{the matrix}

\item{r}{the rows of desired cut.}

\item{c}{the columns of desired cut.}
}
\value{
the splitted matrix
}
\description{
Splits a big matrix in sub matrices using Kronecker product
}
\keyword{internal}
