% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_HVAR.R
\name{HDGC_HVAR}
\alias{HDGC_HVAR}
\title{Test Granger causality in High Dimensional HVARs}
\usage{
HDGC_HVAR(
  GCpair,
  data,
  log = TRUE,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{GCpair}{a named list with names GCto and GCfrom containing vectors of the relevant GC variables.}

\item{data}{the data matrix or an object that can be transformed to a matrix containing (stationary) realized volatilities.}

\item{log}{default is TRUE, if the realized volatilities are already log transformed then put =FALSE}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM Chi-square test statistics (asymptotic), LM F-stat with finite sample correction, LM Chi-square (asymptotic) with heteroscedasticity correction, all with their corresponding p-value.
Lasso selections are also printed to the console.
}
\description{
Test Granger causality in High Dimensional (Stationary) Heterogeneous VARs
}
\examples{
HDGC_HVAR(GCpair=list("GCto"="Var 1", "GCfrom"="Var 5"), data=sample_RV,log=TRUE)
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).

Corsi, Fulvio. "A simple approximate long-memory model of realized volatility." Journal of Financial Econometrics 7.2 (2009): 174-196.
}
