% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luca_Stephan_pkg.R
\name{HDGC_HVAR_multiple_RVCOV}
\alias{HDGC_HVAR_multiple_RVCOV}
\title{Test multiple combinations Granger causality for realized volatilities in High Dimensional Stationary HVARs}
\usage{
HDGC_HVAR_multiple_RVCOV(
  realized_variances,
  realized_correlations,
  GCpairs,
  log = TRUE,
  bound = 0.5 * nrow(realized_variances),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{realized_variances}{Dataset of realized volatilities. A matrix or something that can be coerced to a matrix. Note: the volatilities must not be in logs.}

\item{realized_correlations}{Dataset of realized correlations. To compute realized correlations from realized variances and realized covariances use \code{\link{Realized_corr}}}

\item{GCpairs}{it should contain a nested list. The outer list is all the pairs to be considered. See \code{ \link{HDGC_HVAR_RVCOV}}.
The inner list contains the GCto and GCfrom vectors needed for \code{\link{HDGC_HVAR_RVCOV}}.}

\item{log}{default is TRUE, if the realized volatilities are already log transformed then put to FALSE}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM test statistics, p-values (asymptotic and with finite sample correction) and Lasso selections are printed to the console
}
\description{
This function is a wrapper around \code{\link{HDGC_HVAR_RVCOV}} that allows for multiple combinations to be tested
}
\examples{
\dontrun{ HDGC_HVAR_multiple_RVCOV(real_var, real_corr, GCpairs, log = T,parallel = T)}
}
