% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_VAR_all.R
\name{HDGC_VAR_all}
\alias{HDGC_VAR_all}
\title{Granger Causality Network in High Dimensional mixed Integrated and Cointegrated VARs}
\usage{
HDGC_VAR_all(
  data,
  p = 1,
  d = 0,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or something that can be coerced to a matrix.}

\item{p}{lag length of VAR}

\item{d}{order of lag augmentation corresponding to suspected max order of integration}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
Granger causality matrix and Lasso selections are printed to the console
}
\description{
Wrapper around \code{\link{HDGC_VAR_multiple}} which tests Granger causality from each variable to all other variables,
one by one. Can therefore be used to construct a network.
}
\examples{
\dontrun{HDGC_VAR_all(data, p=2, d=2,parallel = T ) }
}
