% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_VAR_all_I0.R
\name{HDGC_VAR_all_I0}
\alias{HDGC_VAR_all_I0}
\title{Granger Causality Network in High Dimensional Stationary VARs}
\usage{
HDGC_VAR_all_I0(
  data,
  p = 1,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or object that can be coerced to a matrix.}

\item{p}{lag length of VAR}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
Granger causality matrix and Lasso selections are printed to the console
}
\description{
Wrapper around \code{\link{HDGC_VAR_multiple_I0}} which tests Granger causality from each variable to all other variables,
one by one. Can therefore be used to construct a network.
}
\examples{
\dontrun{HDGC_VAR_all_I0(data=sample_dataset_I0,p=2,parallel=T)}
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).
}
