% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_VAR_I0.R
\name{HDGC_VAR_I0}
\alias{HDGC_VAR_I0}
\title{Test Granger causality in High Dimensional Stationary VARs}
\usage{
HDGC_VAR_I0(
  GCpair,
  data,
  p = 1,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{GCpair}{a named list with names GCto and GCfrom containing vectors of the relevant GC variables.}

\item{data}{data matrix or something that can be coerced to a matrix.}

\item{p}{lag length of the VAR}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM test statistics, p values: asymptotic and with finite sample correction and Lasso selections are printed to the console
}
\description{
Test Granger causality in High Dimensional Stationary VARs
}
\examples{
HDGC_VAR_I0(GCpair=list("GCto"="Var 1", "GCfrom"="Var 2"), data=sample_dataset_I0, p=2)
}
