% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_VAR_multiple.R
\name{HDGC_VAR_multiple}
\alias{HDGC_VAR_multiple}
\title{Test multiple combinations Granger causality in High Dimensional mixed Integrated and Cointegrated VARs}
\usage{
HDGC_VAR_multiple(
  data,
  GCpairs,
  p = 1,
  d = 0,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or object that can be coerced to a matrix.}

\item{GCpairs}{it should contain a nested list. The outer list is all the pairs to be considered.
The inner list contains the GCto and GCfrom vectors needed for \code{\link{HDGC_VAR}}.}

\item{p}{lag length of the VAR}

\item{d}{order of lag augmentation corresponding to suspected max order of integration}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM Chi-square test statistics (asymptotic), LM F-stat with finite sample correction, both with their corresponding p-value.
Lasso selections are also printed to the console.
}
\description{
This function is a wrapper around \code{\link{HDGC_VAR}} that allows for multiple combinations to be tested
}
\examples{
\dontrun{GC=list(list("GCto"="Var 1","GCfrom"="Var 2"),list("GCto"="Var 2","GCfrom"="Var 3"))}
\dontrun{HDGC_VAR_multiple(sample_dataset_I1,GCpairs=GC,3,2)}
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Inference in Non Stationary High Dimensional VARs" (2020, check the latest version at https://sites.google.com/view/luca-margaritella )

Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).
}
