% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGC_VAR_multiple_I0.R
\name{HDGC_VAR_multiple_I0}
\alias{HDGC_VAR_multiple_I0}
\title{Test multiple combinations Granger causality in High Dimensional Stationary VARs}
\usage{
HDGC_VAR_multiple_I0(
  data,
  GCpairs,
  p = 1,
  bound = 0.5 * nrow(data),
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{data}{the data matrix or object that can be coerced to a matrix.}

\item{GCpairs}{it should contain a nested list. The outer list is all the pairs to be considered.
The inner list contains the GCto and GCfrom vectors needed for \code{\link{HDGC_VAR_I0}}.}

\item{p}{lag length of the VAR}

\item{bound}{lower bound on tuning parameter lambda}

\item{parallel}{TRUE for parallel computing}

\item{n_cores}{nr of cores to use in parallel computing, default is all but one}
}
\value{
LM Chi-square test statistics (asymptotic), LM F-stat with finite sample correction, both with their corresponding p-value.
Lasso selections are also printed to the console.
}
\description{
This function is a wrapper around \code{\link{HDGC_VAR_I0}} that allows for multiple combinations to be tested
}
\examples{
\dontrun{GC=list(list("GCto"="Var 1", "GCfrom"="Var 2"),list("GCto"="Var 2", "GCfrom"="Var 3"))}
\dontrun{HDGC_VAR_multiple_I0(sample_dataset_I0, GC, p=1 )}
}
