% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_sqrt.R
\name{matrix_sqrt}
\alias{matrix_sqrt}
\title{Calculate square root of a matrix}
\usage{
matrix_sqrt(M)
}
\arguments{
\item{M}{the matrix}
}
\value{
the matrix square rooted
}
\description{
Calculate square root of a matrix
}
\keyword{internal}
