% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GC_all.R
\name{Plot_GC_all}
\alias{Plot_GC_all}
\title{Plot High-Dimensional Granger causality Networks}
\usage{
Plot_GC_all(
  Comb,
  Stat_type = "FS_cor",
  alpha = 0.01,
  multip_corr = list(F, "bonferroni", gamm = c(1e-04, 0.1, 0.001), fdr.apf = c(0.05,
    0.6), verb = F),
  ...,
  cluster = list(F, 10, "black", 0.51, 1, 0)
)
}
\arguments{
\item{Comb}{output from: \code{\link{HDGC_VAR_all_I0}}, \code{\link{HDGC_VAR_multiple_pairs_I0}}, \code{\link{HDGC_VAR_all}}, \code{\link{HDGC_VAR_multiple_pairs}}, \code{\link{HDGC_HVAR_all}} or \code{\link{HDGC_HVAR_multiple_pairs}}}

\item{Stat_type}{either FS_cor (default), Asymp or Asymp_Robust respectively for F-stat small sample correction, standard Chi square test, standard chi square test with heteroscedasticity correction}

\item{alpha}{the desired probability of type one error, default is 0.01.}

\item{multip_corr}{A list: first element is logical, if TRUE a multiple testing correction using \code{\link[stats]{p.adjust}} is used. The second
element of the list define the p.adjust.method=c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", "none")). If the second element gets the name "APF_FDR" then \code{\link[APFr]{apf_fdr}} is called which uses empirical Bayes is called and a third and fourth
elements in the mutip_corr list are required: gamm=c(a,b,c) requires a min (a), max (b) and step length (c) values to be set for the threshold on the p_values,
\code{\link[APFr]{apf_fdr}} requires one or two values: either (NULL,value) or (value,NULL) if one wants to have specified amount of average power (fdr) no matter fdr (average power).
If both (value,value) are given, the calculated threshold will find the closest combination to both apf and fdr desired. The last element
of the list is logical: verbose=TRUE if one wants to know how much apf/fdr the testing has.}

\item{...}{all parameters for the network plot: see example and \code{\link[igraph]{graph_from_adjacency_matrix}} documentation.}

\item{cluster}{A list: first element is logical, if TRUE a cluster plot using \code{\link[igraph]{cluster_edge_betweenness}} is plotted.
Other elements are respectively: vertex.size, vertex.label.color,vertex.label.cex, vertex.label.dist, edge.curved (see \code{\link[igraph]{graph_from_adjacency_matrix}} for details).}
}
\value{
a \code{\link[igraph]{graph_from_adjacency_matrix}} network
}
\description{
Plot High-Dimensional Granger causality Networks
}
\examples{
\dontrun{Plot_GC_all(Comb, "FS_cor",alpha=0.01,multip_corr=list(F), directed=T, layout.circle}
}
\references{
Hecq, A., Margaritella, L., Smeekes, S., "Inference in Non Stationary High Dimensional VARs" (2020, check the latest version at https://sites.google.com/view/luca-margaritella )

Hecq, A., Margaritella, L., Smeekes, S., "Granger Causality Testing in High-Dimensional VARs: a Post-Double-Selection Procedure." arXiv preprint arXiv:1902.10991 (2019).

Newman, Mark EJ, and Michelle Girvan. "Finding and evaluating community structure in networks." Physical review E 69.2 (2004): 026113.

Quatto, Piero, et al. "Brain networks construction using Bayes FDR and average power function." Statistical Methods in Medical Research 29.3 (2020): 866-878.
}
