% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luca_Stephan_pkg.R, R/simplify_list_RV.R
\name{simplify_list_RV}
\alias{simplify_list_RV}
\title{simplify output}
\usage{
simplify_list_RV(test_list, GCpairs)

simplify_list_RV(test_list, GCpairs)
}
\arguments{
\item{test_list}{the test list from  \code{ \link{LM_test_robust}}.}

\item{GCpairs}{the pairs}
}
\value{
the statistics and pvalues

the statistics and pvalues
}
\description{
This function transforms and simplifies output of the GC testing functions below.

This function transforms and simplifies output of the GC testing functions below.
}
